requirements.txt
dj-database-url==0.5.0
Django==2.1.2
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.6
whitenoise==3.2.1
django-crispy-forms==1.7.2
